Finish up setting up the help + add https support + some finetuning to power indices

+++ src/main/java/com/jeno/fantasyleague/backend/model/Contestant.java
@@ -43,7 +43,7 @@ public class Contestant extends UserAudit {
 	@JoinColumn(name = "contestantgroup_id")
 	private ContestantGroup contestant_group;
 
-	private Integer power_index;
+	private Double power_index;
 
 	public Contestant() {
 	}
@@ -96,11 +96,11 @@ public class Contestant extends UserAudit {
 		this.contestant_group = contestant_group;
 	}
 
-	public Integer getPower_index() {
+	public Double getPower_index() {
 		return power_index;
 	}
 
-	public void setPower_index(Integer power_index) {
+	public void setPower_index(Double power_index) {
 		this.power_index = power_index;
 	}
